src/gittest/requirements.txt
# account==0.1.0
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.1
allure-pytest==2.13.5
httpx==0.13.0
allure-python-commons==2.13.5
altgraph==0.17.4
#amqp==2.6.1
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.8.1
asttokens==2.4.1
async-timeout==4.0.3
asyncio==3.4.3  # Если используешь Python 3.6+, можешь убрать
asyncpg==0.29.0
atomicwrites==1.4.1
attrs==23.2.0
bcrypt==4.1.2
billiard==4.2.0
blinker==1.8.0
build==1.2.1
CacheControl==0.14.0
cached-property==1.5.2
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
#chardet==5.2.0
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
crashtest==0.4.1
cryptography==42.0.5
cssselect==1.2.0
cssutils==2.10.2
cv2-brightness==0.0.1
cycler==0.12.1
DAWG-Python==0.7.2
decorator==5.1.1
distlib==0.3.8
distro==1.9.0
dnspython==2.6.1
docker==6.1.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
dulwich==0.21.7
dynaconf==3.2.5
ecdsa==0.19.0
email-validator==1.3.1
#emails==0.5.15
erqa==1.1.2
exceptiongroup==1.2.1 # Если используешь Python 3.11+, можешь убрать
executing==2.0.1
fastapi==0.109.0
fastjsonschema==2.19.1
filelock==3.13.4
Flask==3.0.3
fonttools==4.44.3
frozenlist==1.4.1
fsspec==2024.6.1
future==0.18.3  # Если используешь Python 3+, можешь убрать
gevent==24.2.1
greenlet==3.0.3
gunicorn==20.1.0
h11==0.9.0
h2==3.1.0
hiku==0.7.2
hpack==3.0.0
Hypercorn==0.16.0
hyperframe==5.2.0
#idna==3.7
imageio==2.34.2
importlib_metadata==7.1.0  # Если используешь Python 3.8+, можешь убрать
inflect==5.6.2
inflection==0.5.1
iniconfig==2.0.0
installer==0.7.0
intel-openmp==2021.4.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jedi==0.19.1
Jinja2==3.1.3
jsonschema==3.2.0
jwt==1.3.1
# keyboard==0.13.5  -- Убран, не нужен для Linux, если нет специфической необходимости
keyring==24.3.1
kiwisolver==1.4.5
# kombu==5.3.3   Обновлен, должен быть совместим с celery 5.3.3
lazy_loader==0.4
loguru==0.7.2
lxml==5.2.1
Mako==1.3.3
MarkupSafe==2.1.5
matplotlib==3.8.2
matplotlib-inline==0.1.7
measuring-tools==0.6.0
mkl==2021.4.0
more-itertools==10.2.0
# mouse==0.7.1 -- Убран, не нужен для Linux, если нет специфической необходимости
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
mypy==1.9.0
mypy-extensions==1.0.0
networkx==3.3
numpy==1.26.0
openai==0.28.1
opencv-python==4.10.0.84
packaging==23.2
pandas==2.2.0
paramiko==3.4.0
parso==0.8.4
passlib==1.7.4
peewee==3.17.0
# pefile==2023.2.7 -- Убран, не нужен в Linux
pewee==0.1.0.dev9
pexpect==4.9.0
# pidantic==0.1.2
Pillow==10.1.0
pkginfo==1.10.0
platformdirs==4.2.0
pluggy==1.5.0
poetry==1.8.2
poetry-core==1.9.0
poetry-plugin-export==1.7.1
premailer==3.10.0
# prettytable==3.9.1   Заменено на фиксированную версию
priority==2.0.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyasn1==0.6.0
pycparser==2.22
pydantic==1.10.15
pydantic_core==2.18.1
Pygments==2.17.2
pyinstaller==6.6.0
pyinstaller-hooks-contrib==2024.6
PyJWT==2.9.0
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
PyNaCl==1.5.0
pyparsing==3.1.1
pyproject-api==1.6.1
pyproject_hooks==1.0.0
# PyQt5==5.15.11 -- Убраны Qt-зависимости
# PyQt5-Qt5==5.15.2
# PyQt5_sip==12.15.0
# pyqtgraph==0.13.7
pyrsistent==0.20.0
# PySimpleGUI==4.60.5.0 -- Убран
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-html==4.1.1
pytest-xdist==1.33.0
pytest-metadata==3.1.1
pytest-random-order==1.1.1
python-dateutil==2.8.2
python-dotenv==0.21.1
python-jose==3.3.0
python-multipart==0.0.5
python-openc==1.0.0.14
pytz==2023.4
PyYAML==5.3.1
Quart==0.19.6
rapidfuzz==3.8.1
raven==6.10.0
requests==2.31.0
requests-toolbelt==1.0.0
rsa==4.9
ruff==0.3.7
scikit-image==0.24.0
scipy==1.14.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SoundCard==0.4.2
soundfile==0.12.1
#sql-alchemy-db==0.0.4
SQLAlchemy==1.4.52
sqlalchemy-orm==1.2.10
stack-data==0.6.3
starlette==0.35.1
supervisor==4.2.5
sympy==1.12.1
taskgroup==0.0.0a4
tbb==2021.13.0
tenacity==6.3.1
texttable==1.7.0
tifffile==2024.6.18
tomli==2.0.1
tomlkit==0.12.4
torch==2.3.1
torchaudio==2.3.1
torchvision==0.18.1
#tox==4.15.0
tqdm==4.66.6
traitlets==5.14.3
trove-classifiers==2024.4.10
types-six==1.16.21.20240311  #  Если используешь Python 3.10+, можешь убрать.

typing==3.7.4.3   # Если используешь Python 3.7+, можешь убрать.
typing-inspect==0.9.0
typing_extensions==4.11.0  # Если используешь Python 3.8+, можешь убрать.
tzdata==2023.4
unicorn==2.0.1.post1
urllib3==2.2.1
uvicorn==0.25.0
vine==5.0.0
virtualenv==20.25.1
wcwidth==0.2.13
websocket-client==0.59.0
websockets==12.0
Werkzeug==3.0.2
# win32-setctime==1.1.0 -- Убран, не нужен в Linux
wsproto==1.2.0
yarl==1.9.4
zipp==3.18.1
zope.event==5.0
zope.interface==6.3
celery[redis]<6.0.0,>=5.3.0 #  Указан celery с совместимой версией kombu и redis в качестве брокера сообщений
